+++ packages/shared-contexts/src/contexts/apiKeyContext.tsx
@@ -1,4 +1,4 @@
-import { createContext, useReducer, useMemo } from 'react';
+import React, { createContext, useReducer, useMemo } from 'react';
 
 type Action =
   | {