+++ packages/swr-mock/src/mockMiddlewareFactory.ts
@@ -2,6 +2,7 @@
  * TODO: remove the inline ESLint comment
  * local:ergo/rosen-bridge/ui#441
  */
+
 /* eslint-disable */
 import { Key, Middleware, SWRHook } from 'swr';