+++ wallets/nautilus-wallet/src/global.d.ts
@@ -4,12 +4,14 @@ import { EipWalletApi } from '@rosen-ui/wallet-api';
  * global type augmentation for nautilus wallet
  */
 declare global {
-  let ergoConnector: {
-    nautilus: {
-      connect: (params: { createErgoObject: boolean }) => Promise<boolean>;
-      getContext: () => Promise<EipWalletApi>;
+  interface Window {
+    ergoConnector: {
+      nautilus: {
+        connect: (params: { createErgoObject: boolean }) => Promise<boolean>;
+        getContext: () => Promise<EipWalletApi>;
+      };
     };
-  };
+  }
 }
 
 export {};