+++ wallets/okx-wallet/src/getAddressCreator.ts
@@ -1,0 +1,5 @@
+export const getAddressCreator = () => (): Promise<string> => {
+  return window.okxwallet.bitcoin
+    .getAccounts()
+    .then((accounts: string[]) => accounts[0]);
+};