+++ packages/swr-mock/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rosen-ui/swr-mock
 
+## 0.0.3
+
+### Patch Changes
+
+- Update the source code to address the issues identified by the new ESLint rules configuration
+
 ## 0.0.2
 
 ### Patch Changes